requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
google-api-python-client==2.108.0
newspaper3k==0.2.8
trafilatura==1.6.1
openai==1.3.5
pydantic==2.5.2
requests==2.31.0
tiktoken==0.5.2 